+++ C#/Params.cs
@@ -1,0 +1,8 @@
+namespace CSharpFftDemo
+{
+    public static class Params
+    {
+        internal const int Log2FftSize = 12;
+        internal const int FftRepeat = 1000;
+    }
+}
