+++ C#/packaging/DEBIAN/control
@@ -1,5 +1,5 @@
 Package: fft-benchmark
-Version: 1.2.1
+Version: 1.3
 Section: base
 Priority: optional
 Architecture: armhf
